COR: CADRAN: placement spin_nb_decimal
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@4279 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/init_db.sql
@@ -518,7 +518,7 @@ CREATE TABLE IF NOT EXISTS `syns_cadrans` (
   `posx` int(11) NOT NULL DEFAULT '0',
   `posy` int(11) NOT NULL DEFAULT '0',
   `type` int(11) NOT NULL DEFAULT '0',
-  `angle` float NOT NULL DEFAULT '0',
+  `angle` int(11) NOT NULL DEFAULT '0',
   `fleche_left` tinyint(1) NOT NULL DEFAULT '0',
   `nb_decimal` int(11) NOT NULL DEFAULT '2',
   PRIMARY KEY (`id`),
